tehty muutos

+++ teht_4/teht_4.2.py
@@ -2,12 +2,12 @@
 # käyttäjä antaa negatiivisen tuumamäärän. Sen jälkeen ohjelma lopettaa toimintansa.
 # 1 tuuma = 2.54 centimeters
 
-tuumat = int(input("Anna tuumien määrä: "))
+tuumat = float(input("Anna tuumien määrä: "))
 
 while tuumat >= 0:
     cm = 2.54 * tuumat
     print(f"{tuumat} tuumaa on = {cm}cm")
-    tuumat = int(input("Anna tuumien määrä: "))
+    tuumat = float(input("Anna tuumien määrä: "))
 print("Virheellinen luku, käytä vain positiivisia lukuja")
